+++ AlchemistCode/Cards/Uncommon/Hone.cs
@@ -2,26 +2,20 @@ using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
-using MegaCrit.Sts2.Core.HoverTips;
-using MegaCrit.Sts2.Core.Models.Powers;
 using MegaCrit.Sts2.Core.Nodes.CommonUi;
 
 namespace Alchemist.AlchemistCode.Cards.Uncommon;
 
 public class Hone : AlchemistCard
 {
-    public Hone() : base(1, CardType.Skill, CardRarity.Uncommon, TargetType.Self)
+    public Hone() : base(2, CardType.Skill, CardRarity.Uncommon, TargetType.Self)
     {
+        WithCostUpgradeBy(-1);
         WithKeyword(CardKeyword.Exhaust);
-        WithTips(card => card.IsUpgraded
-            ? new[] { HoverTipFactory.FromPower<StrengthPower>() }
-            : []);
     }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
-        if (IsUpgraded)
-            await PowerCmd.Apply<StrengthPower>(choiceContext, Owner.Creature, 1, Owner.Creature, this);
         var hand = PileType.Hand.GetPile(Owner).Cards
             .Where(c => !c.IsUpgraded).ToList();
         if (hand.Count > 0)